Mlops/anime recommender/requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-text-splitters
chromadb
pandas
python-dotenv
sentence-transformers
flask
prometheus-client
gunicorn